Revert "create top level DukeDS file" in favor of package
This reverts commit fda52181fa7b2ca98175ddd99be8f22b5b7673ed.
