+++ ddsc/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+from ddsc.sdk.dukeds import DukeDS, Session, ItemNotFound, DuplicateNameError
+__all__ = ['DukeDS', 'Session', 'ItemNotFound', 'DuplicateNameError']
