--- ddsc/sdk/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import absolute_import
-from ddsc.sdk.session import Session, ItemNotFound, DuplicateNameError
-__all__ = ['Session', 'ItemNotFound', 'DuplicateNameError']
