+++ ddsc/sdk/tests/test_dukeds.py
@@ -1,6 +1,5 @@
 from unittest import TestCase
-import DukeDS
-from ddsc.sdk import ItemNotFound, DuplicateNameError
+from ddsc import DukeDS, ItemNotFound, DuplicateNameError
 from mock import patch, Mock
 
 
@@ -11,7 +10,7 @@ class TestDukeDS(TestCase):
         self.mouse_rna_project = Mock()
         self.mouse_rna_project.name = 'mouse_rna'
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_list_projects(self, mock_client):
         mock_client.return_value.get_projects.return_value = [
             self.bat_dna_project,
@@ -22,13 +21,13 @@ class TestDukeDS(TestCase):
 
         self.assertEqual(projects, ['bat_dna', 'mouse_rna'])
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_create_project(self, mock_client):
         DukeDS.create_project('mouse', 'Mouse research project')
 
         mock_client.return_value.create_project.assert_called_with('mouse', 'Mouse research project')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_create_project_raises_for_duplicate_project_name(self, mock_client):
         mock_existing_project = Mock()
         mock_existing_project.name = 'mouse'
@@ -37,7 +36,7 @@ class TestDukeDS(TestCase):
         with self.assertRaises(DuplicateNameError):
             DukeDS.create_project('mouse', 'Mouse research project')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_delete_project_not_found(self, mock_client):
         mock_client.return_value.get_projects.return_value = [
             self.bat_dna_project
@@ -45,7 +44,7 @@ class TestDukeDS(TestCase):
         with self.assertRaises(ItemNotFound):
             DukeDS.delete_project('mouse_rna')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_delete_project(self, mock_client):
         mock_client.return_value.get_projects.return_value = [
             self.bat_dna_project,
@@ -57,8 +56,8 @@ class TestDukeDS(TestCase):
         self.bat_dna_project.delete.assert_not_called()
         self.mouse_rna_project.delete.assert_called()
 
-    @patch('ddsc.sdk.session.Client')
-    @patch('ddsc.sdk.session.PathToFiles')
+    @patch('ddsc.sdk.dukeds.Client')
+    @patch('ddsc.sdk.dukeds.PathToFiles')
     def test_list_files(self, mock_path_to_files, mock_client):
         mock_client.return_value.get_projects.return_value = [
             self.bat_dna_project,
@@ -78,7 +77,7 @@ class TestDukeDS(TestCase):
             'data/file2.dat'
         ])
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_download_file(self, mock_client):
         self.mouse_rna_project.get_children.return_value = []
         mock_client.return_value.get_projects.return_value = [
@@ -87,7 +86,7 @@ class TestDukeDS(TestCase):
 
         DukeDS.download_file('mouse_rna', 'data/file1.dat')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_upload_file(self, mock_client):
         self.mouse_rna_project.get_children.return_value = []
         mock_client.return_value.get_projects.return_value = [
@@ -98,7 +97,7 @@ class TestDukeDS(TestCase):
 
         self.mouse_rna_project.upload_file.assert_called_with('/tmp/file1.dat', remote_filename='file1.dat')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_upload_file_creating_project_and_parent_dir(self, mock_client):
         mock_client.return_value.get_projects.return_value = [
         ]
@@ -115,7 +114,7 @@ class TestDukeDS(TestCase):
         mock_project.create_folder.assert_called_with('data')
         mock_folder.upload_file.assert_called_with('/tmp/file1.dat', remote_filename='file1.dat')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_upload_file_makes_new_version(self, mock_client):
         self.mouse_rna_project.get_children.return_value = []
         mock_client.return_value.get_projects.return_value = [
@@ -131,7 +130,7 @@ class TestDukeDS(TestCase):
         self.mouse_rna_project.upload_file.assert_not_called()
         mock_file1.upload_new_version.assert_called_with('/tmp/file1.dat')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_upload_file_default_dest_is_local_basename(self, mock_client):
         self.mouse_rna_project.get_children.return_value = []
         mock_client.return_value.get_projects.return_value = [
@@ -142,7 +141,7 @@ class TestDukeDS(TestCase):
 
         self.mouse_rna_project.upload_file.assert_called_with('/tmp/file1.dat', remote_filename='file1.dat')
 
-    @patch('ddsc.sdk.session.Client')
+    @patch('ddsc.sdk.dukeds.Client')
     def test_delete_file(self, mock_client):
         mock_client.return_value.get_projects.return_value = [
             self.bat_dna_project,
